backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "apscheduler>=3.10.4",
    "asyncpg>=0.31.0",
    "black>=26.3.1",
    "bs4>=0.0.2",
    "fastapi>=0.136.0",
    "google-genai>=1.73.1",
    "httpx>=0.28.1",
    "playwright>=1.58.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
skip-string-normalization = false

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true

[tool.ruff]
line-length = 88

fix = true
